+++ src/Sylius/Component/Addressing/Model/Country.php
@@ -13,6 +13,7 @@ namespace Sylius\Component\Addressing\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Sylius\Component\Resource\Model\ToggleableTrait;
 
 /**
  * @author Paweł Jędrzejewski <pjedrzejewski@sylius.pl>
@@ -21,6 +22,8 @@ use Doctrine\Common\Collections\Collection;
  */
 class Country implements CountryInterface
 {
+    use ToggleableTrait;
+
     /**
      * @var mixed
      */
@@ -37,11 +40,6 @@ class Country implements CountryInterface
      * @var Collection|ProvinceInterface[]
      */
     protected $provinces;
-
-    /**
-     * @var bool
-     */
-    protected $enabled = true;
 
     public function __construct()
     {
@@ -132,37 +130,5 @@ class Country implements CountryInterface
     public function hasProvince(ProvinceInterface $province)
     {
         return $this->provinces->contains($province);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isEnabled()
-    {
-        return $this->enabled;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setEnabled($enabled)
-    {
-        $this->enabled = (bool) $enabled;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function enable()
-    {
-        $this->enabled = true;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function disable()
-    {
-        $this->enabled = false;
     }
 }