+++ src/Sylius/Component/Rbac/Model/PermissionInterface.php
@@ -14,6 +14,7 @@ namespace Sylius\Component\Rbac\Model;
 use Doctrine\Common\Collections\Collection;
 use Sylius\Component\Resource\Model\CodeAwareInterface;
 use Sylius\Component\Resource\Model\ResourceInterface;
+use Sylius\Component\Resource\Model\SoftDeletableInterface;
 use Sylius\Component\Resource\Model\TimestampableInterface;
 
 /**
@@ -21,7 +22,7 @@ use Sylius\Component\Resource\Model\TimestampableInterface;
  *
  * @author Paweł Jędrzejewski <pawel@sylius.org>
  */
-interface PermissionInterface extends CodeAwareInterface, TimestampableInterface, ResourceInterface
+interface PermissionInterface extends CodeAwareInterface, SoftDeletableInterface, TimestampableInterface, ResourceInterface
 {
     /**
      * @return string
@@ -44,12 +45,12 @@ interface PermissionInterface extends CodeAwareInterface, TimestampableInterface
     public function setParent(PermissionInterface $permission);
 
     /**
-     * @return Collection
+     * @return Collection|PermissionInterface[]
      */
     public function getChildren();
 
     /**
-     * @return boolean
+     * @return bool
      */
     public function hasChildren();
 
@@ -66,7 +67,7 @@ interface PermissionInterface extends CodeAwareInterface, TimestampableInterface
     /**
      * @param PermissionInterface $permission
      *
-     * @return boolean
+     * @return bool
      */
     public function hasChild(PermissionInterface $permission);