+++ src/Sylius/Component/Shipping/Model/ShippingMethodInterface.php
@@ -20,7 +20,7 @@ use Sylius\Component\Resource\Model\ToggleableInterface;
  * @author Paweł Jędrzejewski <pawel@sylius.org>
  * @author Gonzalo Vilaseca <gvilaseca@reiss.co.uk>
  */
-interface ShippingMethodInterface extends CodeAwareInterface, TimestampableInterface, ShippingMethodTranslationInterface, ToggleableInterface
+interface ShippingMethodInterface extends CodeAwareInterface, ShippingMethodTranslationInterface, TimestampableInterface, ToggleableInterface
 {
     // Shippables requirement to match given method.
     const CATEGORY_REQUIREMENT_MATCH_NONE = 0;