+++ src/Sylius/Component/User/Model/GroupableInterface.php
@@ -13,6 +13,8 @@
 
 namespace Sylius\Component\User\Model;
 
+use Doctrine\Common\Collections\Collection;
+
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
  * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
@@ -22,14 +24,14 @@ interface GroupableInterface
     /**
      * Gets the groups granted to the user.
      *
-     * @return \Traversable
+     * @return Collection|GroupInterface[]
      */
     public function getGroups();
 
     /**
      * Gets the name of the groups which includes the user.
      *
-     * @return array
+     * @return string[]
      */
     public function getGroupNames();
 
@@ -38,7 +40,7 @@ interface GroupableInterface
      *
      * @param string $name Name of the group
      *
-     * @return Boolean
+     * @return bool
      */
     public function hasGroup($name);