+++ src/Sylius/Component/User/Model/User.php
@@ -15,6 +15,9 @@ namespace Sylius\Component\User\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Sylius\Component\Resource\Model\SoftDeletableTrait;
+use Sylius\Component\Resource\Model\TimestampableTrait;
+use Sylius\Component\Resource\Model\ToggleableTrait;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
@@ -23,6 +26,8 @@ use Doctrine\Common\Collections\Collection;
  */
 class User implements UserInterface
 {
+    use SoftDeletableTrait, TimestampableTrait, ToggleableTrait;
+
     /**
      * @var mixed
      */
@@ -44,11 +49,6 @@ class User implements UserInterface
      * @var string
      */
     protected $usernameCanonical;
-
-    /**
-     * @var boolean
-     */
-    protected $enabled = false;
 
     /**
      * Random data that is used as an additional input to a function that hashes a password.
@@ -111,30 +111,18 @@ class User implements UserInterface
     protected $roles = array(UserInterface::DEFAULT_ROLE);
 
     /**
-     * @var Collection
+     * @var Collection|UserOAuth[]
      */
     protected $oauthAccounts;
-
-    /**
-     * @var \DateTime
-     */
-    protected $createdAt;
-
-    /**
-     * @var \DateTime
-     */
-    protected $updatedAt;
-
-    /**
-     * @var \DateTime
-     */
-    protected $deletedAt;
 
     public function __construct()
     {
         $this->salt = base_convert(sha1(uniqid(mt_rand(), true)), 16, 36);
         $this->oauthAccounts = new ArrayCollection();
         $this->createdAt = new \DateTime();
+
+        // Set here to overwrite default value from trait
+        $this->enabled = false;
     }
 
     /**
@@ -311,22 +299,6 @@ class User implements UserInterface
     /**
      * {@inheritdoc}
      */
-    public function isEnabled()
-    {
-        return $this->enabled;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setEnabled($enabled)
-    {
-        $this->enabled = (bool) $enabled;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function isAccountNonExpired()
     {
         return !$this->hasExpired($this->expiresAt);
@@ -435,7 +407,7 @@ class User implements UserInterface
      */
     public function isPasswordRequestNonExpired(\DateInterval $ttl)
     {
-        return (null !== $this->passwordRequestedAt && new \DateTime() <= $this->passwordRequestedAt->add($ttl));
+        return null !== $this->passwordRequestedAt && new \DateTime() <= $this->passwordRequestedAt->add($ttl);
     }
 
     /**
@@ -504,62 +476,6 @@ class User implements UserInterface
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function getCreatedAt()
-    {
-        return $this->createdAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCreatedAt(\DateTime $createdAt)
-    {
-        $this->createdAt = $createdAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getUpdatedAt()
-    {
-        return $this->updatedAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setUpdatedAt(\DateTime $updatedAt)
-    {
-        $this->updatedAt = $updatedAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDeletedAt()
-    {
-        return $this->deletedAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setDeletedAt(\DateTime $deletedAt = null)
-    {
-        $this->deletedAt = $deletedAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isDeleted()
-    {
-        return $this->hasExpired($this->deletedAt);
-    }
-
-    /**
      * Returns username.
      *
      * @return string
@@ -605,7 +521,7 @@ class User implements UserInterface
             $this->locked,
             $this->enabled,
             $this->id
-            ) = $data;
+        ) = $data;
     }
 
     /**
@@ -619,10 +535,11 @@ class User implements UserInterface
     }
 
     /**
+     * @param \DateTime $date
      * @return bool
      */
-    protected function hasExpired($date)
+    protected function hasExpired(\DateTime $date = null)
     {
-        return (null !== $date) && ((new \DateTime()) >= $date);
+        return null !== $date && new \DateTime() >= $date;
     }
 }