+++ src/Sylius/Component/User/Model/UserInterface.php
@@ -17,6 +17,7 @@ use Doctrine\Common\Collections\Collection;
 use Sylius\Component\Resource\Model\ResourceInterface;
 use Sylius\Component\Resource\Model\SoftDeletableInterface;
 use Sylius\Component\Resource\Model\TimestampableInterface;
+use Sylius\Component\Resource\Model\ToggleableInterface;
 use Symfony\Component\Security\Core\User\AdvancedUserInterface;
 
 /**
@@ -27,10 +28,11 @@ use Symfony\Component\Security\Core\User\AdvancedUserInterface;
 interface UserInterface extends
     AdvancedUserInterface,
     CredentialsHolderInterface,
+    ResourceInterface,
     \Serializable,
-    TimestampableInterface,
     SoftDeletableInterface,
-    ResourceInterface
+    TimestampableInterface,
+    ToggleableInterface
 {
     const DEFAULT_ROLE = 'ROLE_USER';
 
@@ -40,7 +42,7 @@ interface UserInterface extends
     public function getEmail();
 
     /**
-     * @param  string $email
+     * @param string $email
      */
     public function setEmail($email);
 
@@ -52,7 +54,7 @@ interface UserInterface extends
     public function getEmailCanonical();
 
     /**
-     * @param  string $emailCanonical
+     * @param string $emailCanonical
      */
     public function setEmailCanonical($emailCanonical);
 
@@ -84,12 +86,7 @@ interface UserInterface extends
     public function setUsernameCanonical($usernameCanonical);
 
     /**
-     * @param boolean $enabled
-     */
-    public function setEnabled($enabled);
-
-    /**
-     * @param boolean $locked
+     * @param bool $locked
      */
     public function setLocked($locked);
 
@@ -115,7 +112,7 @@ interface UserInterface extends
      *
      * @param \DateInterval $ttl Requests older than this time interval will be considered expired
      *
-     * @return boolean true if the user's password request is non expired, false otherwise
+     * @return bool true if the user's password request is non expired, false otherwise
      */
     public function isPasswordRequestNonExpired(\DateInterval $ttl);
 
@@ -144,7 +141,7 @@ interface UserInterface extends
      *
      * @param string $role
      *
-     * @return boolean
+     * @return bool
      */
     public function hasRole($role);