+++ src/Sylius/Component/Resource/CHANGELOG.md
@@ -10,3 +10,4 @@ CHANGELOG
 * Extended RepositoryInterface by ``add`` and ``remove`` methods.
 * Introduce Metadata class for easier handling of resource configurations.
 * Add Registry to hold all resources in the system.
+* [BC BREAK] Introduced 3 new traits: `SoftDeletableTrait`, `TimestampableTrait`, `ToggleableTrait`